+++ dash_leaflet/express.py
@@ -1,4 +1,21 @@
+import logging
+
 import dash_leaflet as dl
+import base64
+
+
+def _try_import_geobuf():
+    install_txt = "e.g. via pip by running 'pip install dash-leaflet[geobuf]' or 'pip install dash-leaflet[all]'."
+    try:
+        import geobuf
+        import google.protobuf
+    except ImportError as ex:
+        logging.error(f"Unable to import [geobuf/protobuf]. Please install it, {install_txt}")
+        raise ex
+    if google.protobuf.__version__ != "3.20.0":
+        version_txt = f"The recommended protobuf version is 3.20.0 (you have {google.protobuf.__version__})"
+        logging.warning(f"{version_txt}. You can fix it {install_txt}")
+    return geobuf
 
 
 def categorical_colorbar(*args, categories, colorscale, **kwargs):
@@ -16,3 +33,8 @@ def dicts_to_geojson(dicts, lat="lat", lon="lon"):
             feature["properties"] = {prop: d[prop] for prop in props}
         geojson["features"].append(feature)
     return geojson
+
+
+def geojson_to_geobuf(geojson):
+    geobuf = _try_import_geobuf()
+    return base64.b64encode(geobuf.encode(geojson)).decode()